+++ tests/test_integration.py
@@ -43,14 +43,53 @@ class Test_kmedoids(unittest.TestCase):
         assert np.array_equal(alt.medoids, alt_rust[2])
         assert alt.loss == alt_rust[0]
 
+    def test_pamsil(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.float32)
+        pamsil = kmedoids.pamsil(dist, 2)
+        pamsil_rust = kmedoids.kmedoids._pamsil_swap_f32(dist, pamsil.medoids, 100)
+        assert pamsil.loss == 0.5137878787878788
+        assert pamsil.loss == pamsil_rust[0]
+        assert np.array_equal(pamsil.medoids, pamsil_rust[2])
+
+    def test_pammedsil(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.float32)
+        pms = kmedoids.pammedsil(dist, 2)
+        pms_rust = kmedoids.kmedoids._pammedsil_swap_f32(dist, pms.medoids, 100)
+        assert pms.loss == 0.8172727272727272
+        assert pms.loss == pms_rust[0]
+        assert np.array_equal(pms.medoids, pms_rust[2])
+
+    def test_fastmsc(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.float32)
+        fmsc = kmedoids.fastmsc(dist, 2, init="build")
+        fmsc_rust = kmedoids.kmedoids._fastmsc_f32(dist, fmsc.medoids, 100)
+        assert fmsc.loss == 0.8172727272727272
+        assert np.array_equal(fmsc.medoids, fmsc_rust[2])
+        assert fmsc.loss == fmsc_rust[0]
+
+    def test_fastermsc(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.float32)
+        fmsc = kmedoids.fastermsc(dist, 2, init="build")
+        fmsc_rust = kmedoids.kmedoids._fastermsc_f32(dist, fmsc.medoids, 100)
+        assert fmsc.loss == 0.8172727272727272
+        assert np.array_equal(fmsc.medoids, fmsc_rust[2])
+        assert fmsc.loss == fmsc_rust[0]
+
     def test_silhouette(self):
         dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)
         pam = kmedoids.pam(dist, 2)
-        sil = kmedoids.silhouette(dist, pam.labels)
+        sil = kmedoids.silhouette(dist, pam.labels, n_cpu=1)
         par_sil = kmedoids.silhouette(dist, pam.labels, n_cpu=2)
         sil_rust = kmedoids.kmedoids._silhouette_i32(dist, pam.labels, False)
-        assert sil == par_sil
-        assert sil == sil_rust[0]
+        assert sil[0] == par_sil
+        assert sil[0] == sil_rust[0]
+
+    def test_medoid_silhouette(self):
+        dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.float32)
+        pam = kmedoids.fasterpam(dist, 2)
+        sil = kmedoids.medoid_silhouette(dist, pam.labels)
+        sil_rust = kmedoids.kmedoids._medoid_silhouette_f32(dist, pam.labels, False)
+        assert sil[0] == sil_rust[0]
 
     def test_sklearn_interface(self):
         dist = np.array([[0, 2, 3, 4, 5], [2, 0, 6, 7, 8], [3, 6, 0, 9, 10], [4, 7, 9, 0, 11], [5, 8, 10, 11, 0]], dtype=np.int32)